fix file logging
module_start is called at init, this is before _start is called

for the ps4's orbis sdk, this is roughly the initialization sequence

https://github.com/OpenOrbis/musl/blob/779f95174b44d39e6a8a788b36289cf4768944a9/arch/ps4/crt_arch.h

https://github.com/OpenOrbis/musl/blob/779f95174b44d39e6a8a788b36289cf4768944a9/crt/ps4/crt1.c#L23

https://github.com/OpenOrbis/musl/blob/779f95174b44d39e6a8a788b36289cf4768944a9/src/env/__libc_start_main.c#L79

for now, run everything lazily. this will break when pre-main calls an extern. we want to log pre-main externs, not just code inside main.

I think we want pre-main logging to hit a buffer and post-main we have a thread flush stuff every so often.