use pthread_self based thread local storage
very gross but it seems to not exhaust threads. I am aware that this tls logging does not flush if the thread dies, the way to do this is probably send all tls ring buffers somewhere upon creation and let some flusher inspect all of them. not sure how to do this safely and correctly though so some logs dropped for now