alloc instead of dropping
it seems like we were dropping quite a few messages before. this changes things so we don't drop stuff and instead allocate.

allocating is not wait or lock free but hopefully we don't allocate too often and there is enough space to match the production rate

things will hitch for a second in the worst case but hopefully it will be once and hopefully it will be early

next up test it