add shader collection to extradata
this is a bit jank but the goal here is to make forward progress, not for things to be perfect just yet. we don't know whether this approach is even going to work without looking at the shaders and taking them apart

i'd like to have
1. have the rust code only handle parameter extraction and not deal with all this extra data riff raff
2. reuse the encoder / decoder defs between the viewer and plugin
3. not have the emitting code be so manual
4. idk there were 6 things but i forgot what the other 3 were

point is this is messy. i don't like the wire format right now